Simplify state counting so it only happens once per request
