+++ tests/components/deluge/__init__.py
@@ -23,9 +23,7 @@ GET_TORRENT_STATUS_RESPONSE = {
 }
 
 GET_TORRENT_STATES_RESPONSE = {
-    "states": {
-        "6dcd3f46d09547b62bf07ba9b2943c95d53ddae3": {b"state": b"Seeding"},
-        "1c56ea49918b9baed94cf4bc0ee9f324efc8841a": {b"state": b"Downloading"},
-        "fbf4dab701189a344fa5ab06d7b87c11a74e3da0": {b"state": b"Seeding"},
-    }
+    "6dcd3f46d09547b62bf07ba9b2943c95d53ddae3": {b"state": b"Seeding"},
+    "1c56ea49918b9baed94cf4bc0ee9f324efc8841a": {b"state": b"Downloading"},
+    "fbf4dab701189a344fa5ab06d7b87c11a74e3da0": {b"state": b"Seeding"},
 }
